+++ fixtures/core_test/no_reload.clj
@@ -1,0 +1,4 @@
+(ns ^:clj-reload/no-reload no-reload)
+
+(def rand1
+  (rand-int Integer/MAX_VALUE))