+++ fixtures/core_test/no_unload.clj
@@ -1,0 +1,7 @@
+(ns ^:clj-reload/no-unload no-unload)
+
+(def rand1
+  (rand-int Integer/MAX_VALUE))
+
+(def rand2
+  (rand-int Integer/MAX_VALUE))